refactor: split course commands into different files
