+++ src/utils/operatorParser.ts
